LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/cloning-steps-expected.txt


PASS input element's value should be cloned
PASS input element's dirty value flag should be cloned, so setAttribute doesn't affect the cloned input's value
PASS input[type=button] element's checkedness should be cloned
FAIL input[type=button] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=checkbox] element's checkedness should be cloned
FAIL input[type=checkbox] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=color] element's checkedness should be cloned
FAIL input[type=color] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=date] element's checkedness should be cloned
FAIL input[type=date] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=datetime-local] element's checkedness should be cloned
FAIL input[type=datetime-local] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=email] element's checkedness should be cloned
FAIL input[type=email] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=file] element's checkedness should be cloned
FAIL input[type=file] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=hidden] element's checkedness should be cloned
FAIL input[type=hidden] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=image] element's checkedness should be cloned
FAIL input[type=image] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=month] element's checkedness should be cloned
FAIL input[type=month] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=number] element's checkedness should be cloned
FAIL input[type=number] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=password] element's checkedness should be cloned
FAIL input[type=password] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=radio] element's checkedness should be cloned
FAIL input[type=radio] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=range] element's checkedness should be cloned
FAIL input[type=range] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=reset] element's checkedness should be cloned
FAIL input[type=reset] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=search] element's checkedness should be cloned
FAIL input[type=search] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=submit] element's checkedness should be cloned
FAIL input[type=submit] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=tel] element's checkedness should be cloned
FAIL input[type=tel] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=text] element's checkedness should be cloned
FAIL input[type=text] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=time] element's checkedness should be cloned
FAIL input[type=time] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=url] element's checkedness should be cloned
FAIL input[type=url] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
PASS input[type=week] element's checkedness should be cloned
FAIL input[type=week] element's dirty checkedness should be cloned, so setAttribute doesn't affect the cloned input's checkedness assert_equals: expected false but got true
